+++ NEWS.md
@@ -6,6 +6,7 @@ New features
 
 * New ``pp_cfg`` function (and S3 class) allows one to add a pre-computed cache of component opt to configuration lists (#112). 
 * ``make_pnp`` function now supports adding piecepack pyramids (#37) and matchsticks (#69) to the print-and-play layout.
+* Now exports several additional utility functions for those who would like to use their own custom draw function (#116).  See ``help("pp_utils")``, ``help("draw_fn_helpers")``, and ``help("draw_component")`` for more info.
 
 Breaking changes
 ----------------